[TPDE] FIX: phi handling overwrote register allocations
when the branching logic decided not to spill a register because it is only used
in the next block and the phi-logic needs a temporary register and there is no
free register available, it might decide to spill the previously mentioned register
setting its state as spilled and possibly even using that register for a fixed
assignment. However, when the actual program control flow reaches the next block
the variable is not actually spilled but the code generator thinks it is and
initiates a reload that loads incorrect values.

We fix this by making the scratch allocation in the phi-logic restore
the register state when it is destroyed.

NOTE: DE should also be subject to this bug but likely does not run into it
since the implementation does not use ScratchRegs as values as aggresively as we do
and therefore always has free registers available.

Assembly example (from TPCDS query 18):
```
mov eax, /* some computation result */
// spill registers used outside of the next basic block
jnz next_block
// propagate values to phi of other_block
mov [rbp - 0xYY], eax // spill eax
mov eax, 0
mov [rbp - 0xXX], rax // assign to phi

// next_block:
movsx rax, [rbp - 0xYY] // code thinks eax is spilled
                        // because the phi logic spilled it
                        // but the actual last block did not spill the register
add rax, 1
mov [rbp - 0xXX], rax
// fall-through

// other_block:
mov r11, [rbp - 0xXX] // use phi
// ...
```