[LLVM] Hacky fix for invoke with stack args
An invoke that passes arguments on the stack subtracts some offset from
the stack pointer. In the case of unwinding, this offset is currently
not added again. LLVM emits some special CFI directive for this, but
this appears to be non-standard. Until we evaluate a proper fix for this
problem, fix the current crashes by pretending that an invoke is a
dynamic alloca and force the epilogue to use the frame pointer instead
of the possibly bogus stack pointer.