[TPDE] Don't salvage variable refs
Loading the address of a variable into a register typically amounts to
one or two instructions. As there might be more references to the same
variable, eagerly overwriting the address "just because" causes frequent
recalculation of the same address. Not sure if this is the best idea.

Addresses the code size regression of c12bcf61.