[TPDE] Don't spill value shortly before they die
If a value's liveness ends at the current block, then it's not necessary
to spill the value. Although there might be other users in the block,
this is not relevant for subsequent blocks.

As side effect, this avoids looking at incoming PHI values twice per
branch, which, depending on the IR, can be very expensive (especially
for LLVM, where getIncomingValForBlock has linear runtime).

Fix cases where values were not properly ref-counted in the LLVM
compiler.